ml-service/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
numpy
pymongo
motor
python-dotenv
